requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.1
frozenlist==1.3.3
h11==0.14.0
idna==3.4
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.3
openai==0.27.5
pgvector==0.1.8
Pillow==9.5.0
psycopg2-binary==2.9.6
pydantic==1.10.7
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
python-docx==0.8.11
python-dotenv==1.0.0
redis==4.5.4
regex==2023.3.23
requests==2.29.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.26.1
tiktoken==0.3.3
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
yarl==1.9.2
